Removing unnecessary casts
git-svn-id: https://svn.apache.org/repos/asf/httpcomponents/httpclient/trunk@617207 13f79535-47bb-0310-9956-ffa450edef68

+++ module-client/src/main/java/org/apache/http/impl/auth/DigestScheme.java
@@ -431,8 +431,8 @@ public class DigestScheme extends RFC2617Scheme {
 
         char[] buffer = new char[32];
         for (int i = 0; i < 16; i++) {
-            int low = (int) (binaryData[i] & 0x0f);
-            int high = (int) ((binaryData[i] & 0xf0) >> 4);
+            int low = (binaryData[i] & 0x0f);
+            int high = ((binaryData[i] & 0xf0) >> 4);
             buffer[i * 2] = HEXADECIMAL[high];
             buffer[(i * 2) + 1] = HEXADECIMAL[low];
         }
